Provide a stale_age to mkpidlock
Normally the pidlock should be held for just a few seconds; waiting for
its age to exceed 60 seconds until we consider it stale should be safe.
In fact Julia waits 5 times longer if the process creating the pid lock
file seems to be still running.

On the other hand, without a stale age, the lock file is *never*
considered stale, even if the process creating it definitely is gone,
and so the user can get stuck, which obviously is very bad. To get
unstuck they need to manually delete the lock file.